Refactor test classes to their specific package.
